+++ packages/showcases/react/.storybook/preview.js
@@ -1,0 +1,6 @@
+import backgrounds from './addons/backgrounds.json';
+
+export const parameters = {
+  actions: { argTypesRegex: '^on[A-Z].*' },
+  backgrounds,
+};