+++ packages/sources/css/config/postcss/postcss.config.js
@@ -3,6 +3,7 @@ const postcssColorMod = require('postcss-color-mod-function');
 const tailwindcss = require('tailwindcss');
 const autoprefixer = require('autoprefixer');
 const postcssCombineDuplicatedSelectors = require('postcss-combine-duplicated-selectors');
+const postcssDiscardComments = require('postcss-discard-comments');
 
 module.exports = {
   plugins: [
@@ -11,5 +12,6 @@ module.exports = {
     tailwindcss,
     autoprefixer,
     postcssCombineDuplicatedSelectors,
+    postcssDiscardComments,
   ],
 };