+++ packages/sources/react/src/index.ts
@@ -1,0 +1,1 @@
+export { VtmnButton } from './components/VtmnButton';