+++ packages/sources/css/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.6](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.1.5...@vtmn/css@0.1.6) (2021-02-10)
+
+**Note:** Version bump only for package @vtmn/css
+
+
+
+
+
+## [0.1.5](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.1.4...@vtmn/css@0.1.5) (2021-02-10)
+
+**Note:** Version bump only for package @vtmn/css
+
+
+
+
+
 ## [0.1.4](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.1.3...@vtmn/css@0.1.4) (2021-02-08)
 
 **Note:** Version bump only for package @vtmn/css