+++ packages/sources/web-components/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/web-components@0.2.1...@vtmn/web-components@0.2.2) (2021-02-10)
+
+**Note:** Version bump only for package @vtmn/web-components
+
+
+
+
+
+## [0.2.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/web-components@0.2.0...@vtmn/web-components@0.2.1) (2021-02-10)
+
+**Note:** Version bump only for package @vtmn/web-components
+
+
+
+
+
+# [0.2.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/web-components@0.1.1...@vtmn/web-components@0.2.0) (2021-02-10)
+
+
+### Features
+
+* add React components library ([7905dc3](https://github.com/Decathlon/vitamin-web/commit/7905dc3d1659c5da0e6463aa61352f65cf030ed7))
+
+
+
+
+
 ## [0.1.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/web-components@0.1.0...@vtmn/web-components@0.1.1) (2021-02-08)
 
 **Note:** Version bump only for package @vtmn/web-components